Feat@remove global variables + CRAN fixes (#78)
* Use .naryn environment instead of global variables

* export .naryn evironment

* use .emr_env for calling environment of cpp functions

* export C function names explicitly
added a "C_" prefix for functions with an R function with the same name

* bumped version

* fixed some typos

* Added CRAN badge

* Added CRAN comments

* first submission to CRAN

* Get rid of obsolete ftime

* some fixes for SUN systems

* Use Makevars instead of a Makefile

* Added missing docs + logical track to examples db

* remove non-standard compilation flags

* remove stdout, stderr and exit calls

* remove unused files

* Added some apple specific implementations

* fix: test for SIGTERM when waiting for kid

* registered C routines using: R_registerRoutines and R_useDynamicSymbols

* copy examples db on init
Instead of connecting to the installation path, which may be read-only.
Also - CRAN do not like hidden files in packages

* fixed 'stdout' cran note

* use st_mtimespec instead of st_mtim on MAC

* Do not use MAP_POPULATE on MAC

* use copyfile to copy files on MAC

* fixed compilation on MAC: use unsigned template instead of size_t

* fix some minor compilation errors of using std::string in sprintf

* revert change in NRIdsDist - did not pass regression tests

* changed all size_t to uint64_t

* Do not use reverse iterator in emr_quantiles
again compilation problems in mac
(CLANG problem this time, probably)

* additional minor compilation fixes

* removed unused variables

* fix: crash when overriding a track on MAC

* fix a test in mac
the parsing of the directory might be wrong - but the error is OK

* changed title

* fixed a typo

* Fixed bad link in news

* updated cran-comments

* Fixed CRAN comments

* Changed the example of `emr_traceback()` as suggested by CRAN

* updated CRAN-SUBMISSION file

* updated README

+++ src/EMRIdTimeInterval.cpp
@@ -28,9 +28,9 @@ void EMRIdTimeIntervals::sort_and_unify_overlaps(unsigned stime, unsigned etime)
 
     sort(begin(), end());
 
-    size_t cur_idx = 0;
+    uint64_t cur_idx = 0;
 
-    for (size_t i = 1; i < size(); i++) {
+    for (uint64_t i = 1; i < size(); i++) {
         if (at(cur_idx).id != at(i).id || at(cur_idx).tinterv.etime < at(i).tinterv.stime)
             at(++cur_idx) = at(i);
         // unite overlapping intervals
